requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Scheduler
apscheduler==3.10.4

# Scraping
crawl4ai==0.7.8
beautifulsoup4==4.12.3
lxml==5.1.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Utils
python-json-logger==2.0.7
aiohttp==3.9.1